Case Study 1 - Postgres/requirements.txt
Flask>=2.0
psycopg2-binary>=2.9
python-dotenv>=0.19  # <<< This is for load_dotenv()
requests>=2.25       # For test_app.py
pytest>=6.2          # For test_app.py